inline frozen forward chats hook
